--- src/main/java/frc/robot/hardware/phoenix6/accelerometer/PigeonAccelerometer.java
@@ -1,44 +1,0 @@
-package frc.robot.hardware.phoenix6.accelerometer;
-
-import com.ctre.phoenix6.hardware.ParentDevice;
-import frc.robot.hardware.interfaces.IAccelerometer;
-import frc.robot.hardware.phoenix6.Phoenix6Device;
-import frc.robot.hardware.phoenix6.Pigeon2Wrapper;
-
-public class PigeonAccelerometer extends Phoenix6Device implements IAccelerometer {
-
-	Pigeon2Wrapper accelerometer;
-	String logPath;
-
-	public PigeonAccelerometer(String logPath, Pigeon2Wrapper accelerometer) {
-		super(logPath);
-		this.logPath = logPath;
-		this.accelerometer = accelerometer;
-	}
-
-	@Override
-	public double getAccelerationX() {
-		return accelerometer.getAccelerationX().getValue().magnitude();
-	}
-
-	@Override
-	public double getAccelerationY() {
-		return accelerometer.getAccelerationY().getValue().magnitude();
-	}
-
-	@Override
-	public double getAccelerationZ() {
-		return accelerometer.getAccelerationZ().getValue().magnitude();
-	}
-
-	@Override
-	public void log() {
-		log(logPath);
-	}
-
-	@Override
-	public ParentDevice getDevice() {
-		return accelerometer;
-	}
-
-}
